Implement employee Status component, move reservation Status to appropriate directory
