+++ client/src/components/Reservations/Table.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useMemo, useState } from "react";
 import { Table, TableHeader, TableColumn, TableBody, TableRow, TableCell, User, DatePicker, Input, Spinner, Pagination, Button, useDisclosure } from "@nextui-org/react";
 import { parseDate, parseTime } from "@internationalized/date";
-import Status from "@components/Status/Status";
+import Status from "@components/Status/Reservation/Status";
 import type { ReservationData } from "@core/types";
 import ReservationsActions from "@components/Reservations/Actions";
 import { getFullName } from "@core/utils";
